[NFC] Refactor out subtyping discovery code
